requirements.txt
aiohttp~=3.7.4
aiosignal==1.2.0
async-timeout<4.0
attrs==21.2.0
certifi==2021.10.8
chardet<4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
nhentai.py~=2.2.1
discord.py-message-components==1.7.5.4
Flask==2.0.2
frozenlist==1.2.0
idna==2.5
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
psutil==5.8.0
requests~=2.23.0
saucenao-api==2.4.0
typing-extensions==4.0.1
urllib3==1.25.2
Werkzeug==2.0.2
yarl==1.7.2
